requirements.txt
annotated-types==0.7.0 ; python_version >= "3.13"
anyio==4.11.0 ; python_version >= "3.13"
bcrypt==4.0.1 ; python_version >= "3.13"
certifi==2025.8.3 ; python_version >= "3.13"
click==8.3.0 ; python_version >= "3.13"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version >= "3.13"
dnspython==2.8.0 ; python_version >= "3.13"
ecdsa==0.19.1 ; python_version >= "3.13"
email-validator==2.3.0 ; python_version >= "3.13"
fastapi-cli==0.0.13 ; python_version >= "3.13"
fastapi-cloud-cli==0.2.1 ; python_version >= "3.13"
fastapi==0.117.1 ; python_version >= "3.13"
greenlet==3.2.4 ; python_version == "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.16.0 ; python_version >= "3.13"
httpcore==1.0.9 ; python_version >= "3.13"
httptools==0.6.4 ; python_version >= "3.13"
httpx==0.28.1 ; python_version >= "3.13"
idna==3.10 ; python_version >= "3.13"
itsdangerous==2.2.0 ; python_version >= "3.13"
jinja2==3.1.6 ; python_version >= "3.13"
markdown-it-py==4.0.0 ; python_version >= "3.13"
markupsafe==3.0.2 ; python_version >= "3.13"
mdurl==0.1.2 ; python_version >= "3.13"
orjson==3.11.3 ; python_version >= "3.13"
passlib==1.7.4 ; python_version >= "3.13"
pyasn1==0.6.1 ; python_version >= "3.13"
pydantic-core==2.33.2 ; python_version >= "3.13"
pydantic-extra-types==2.10.5 ; python_version >= "3.13"
pydantic-settings==2.11.0 ; python_version >= "3.13"
pydantic==2.11.9 ; python_version >= "3.13"
pygments==2.19.2 ; python_version >= "3.13"
python-dotenv==1.1.1 ; python_version >= "3.13"
python-jose==3.5.0 ; python_version >= "3.13"
python-multipart==0.0.20 ; python_version >= "3.13"
pyyaml==6.0.3 ; python_version >= "3.13"
rich-toolkit==0.15.1 ; python_version >= "3.13"
rich==14.1.0 ; python_version >= "3.13"
rignore==0.6.4 ; python_version >= "3.13"
rsa==4.2 ; python_version >= "3.13"
sentry-sdk==2.39.0 ; python_version >= "3.13"
shellingham==1.5.4 ; python_version >= "3.13"
six==1.17.0 ; python_version >= "3.13"
sniffio==1.3.1 ; python_version >= "3.13"
sqlalchemy==2.0.43 ; python_version >= "3.13"
starlette==0.48.0 ; python_version >= "3.13"
typer==0.19.2 ; python_version >= "3.13"
typing-extensions==4.15.0 ; python_version >= "3.13"
typing-inspection==0.4.1 ; python_version >= "3.13"
ujson==5.11.0 ; python_version >= "3.13"
urllib3==2.5.0 ; python_version >= "3.13"
uvicorn==0.37.0 ; python_version >= "3.13"
uvloop==0.21.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.13"
watchfiles==1.1.0 ; python_version >= "3.13"
websockets==15.0.1 ; python_version >= "3.13"
